Automate container image build and publish to ECR
Set up GitHub action job to build and publish the autoretrieve container
image to a repository set up in ECR at storetheindex AWS account.

The image tag is determined based on the build trigger: either merge to
 `master` or publication of a GitHub release.

If the build is triggered from merge to `main`, then the tag will follow
`timestamp-<commit-sha>` format. The rationale for this for mat is to
have some monotonically increasing numerical valu in the tag so that the
tags can be sortable by age. This is then used by Flux CD to determine
whether there is a new container image available for automated
deployment.

Otherwise, the release name is expected to be a semver starting with `v`
and the tag will be set to the semantic version stripping off the `v`
prefix.