Do not create a PR in CD branches if there is no commit
The CI job that creates PRs from the `cd/dev` branch is triggered on
push. If a push is made to those branches where the is no commit from
the main branch to create a PR the build will fail.

Check the error message for a case where the CD branch is pushed to but
has no commits ahead of the main branch and gracefully terminate the
build.