feat: confirm retrievals resulted in a block in the blockstore
Upon successful retrieval, provide a confirmation along with the reported event
that the retrieval resulted in the block being stored in the blockstore. Using
`retrievedCids` as a proxy isn't enough because it's possible (with some
non-trivial probability) that a retrieval takes place for a block that's just
been retrieved as part of a parallel retrieval, so graphsync doesn't need to
transfer any blocks.